# openapi.yaml
openapi: 3.1.0
info:
  title: AIDecepticon Control Plane API
  version: 0.1.0
  description: API for managing organization-isolated deception deployments, sensors, tokens, incidents, encrypted secrets, audit events, identity domains, and integrations. Operator endpoints use X-AIDecepticon-Organization to select an authorized organization.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8787/api/v1
security:
  - controlPlaneSession: []
  - controlPlaneBearer: []
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      summary: List organizations available to the authenticated identity
      responses:
        '200':
          description: Accessible organizations and the active organization ID
          content:
            application/json:
              schema:
                type: object
                required: [items, activeOrganizationId]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Organization' }
                  activeOrganizationId: { type: string, pattern: '^org-' }
        '403': { description: Identity has no organization assignment }
    post:
      operationId: createOrganization
      summary: Create an isolated enterprise or managed-customer organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
                slug: { type: string, maxLength: 63 }
                plan: { type: string, enum: [managed, enterprise, evaluation] }
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '403': { description: Organization creation requires platform-wide scope }
        '409': { description: Organization slug already exists }
...
components:
  schemas:
    Organization:
      type: object
      required: [id, name, slug, status, plan, createdAt]
      properties:
        id: { type: string, pattern: '^org-' }
        name: { type: string }
        slug: { type: string }
        status: { type: string, enum: [active, suspended] }
        plan: { type: string }
        createdAt: { type: string, format: date-time }
...
  securitySchemes:
    controlPlaneSession:
      type: apiKey
      in: cookie
      name: __Host-aidecepticon.sid
      description: HTTP-only server-side operator session established through OIDC, SAML, or API-key exchange.
    controlPlaneBearer:
      type: http
      scheme: bearer
      description: Optional control-plane API key configured by the administrator.
    sensorBearer:
      type: http
      scheme: bearer
      description: Unique sensor access token returned once during enrollment.
...